examples/flask_api/requirements.txt
assertpy
behave
behave-restful
bolt-ta
flask
Pillow